Apply some more consistency for deployment tests
This change makes deployment tests more consistent and DRY by applying a
table-driven approach. This will help us in the future as we scale up
the number of these.

In addition to that I added some more logging and robustness. There's
actually a good pattern I found for "unconditional cleanup", and I'm
applying it here for app deletion. The problem that this is solves is
that we want to delete the application even if deployment fails, but we
also want a deletion failure to fail the test.
