+++ pkg/curp/handlers/resource_handler.go
@@ -1,0 +1,31 @@
+// ------------------------------------------------------------
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+// ------------------------------------------------------------
+
+package handlers
+
+import (
+	"context"
+
+	"github.com/Azure/radius/pkg/curp/db"
+	"github.com/Azure/radius/pkg/workloads"
+)
+
+type PutOptions struct {
+	Application string
+	Component   string
+	Resource    workloads.WorkloadResource
+	Existing    *db.DeploymentResource
+}
+
+type DeleteOptions struct {
+	Application string
+	Component   string
+	Existing    db.DeploymentResource
+}
+
+type ResourceHandler interface {
+	Put(ctx context.Context, options PutOptions) (map[string]string, error)
+	Delete(ctx context.Context, options DeleteOptions) error
+}
