+++ pkg/workloads/ingress/decorator.go
@@ -47,7 +47,7 @@ func (r Renderer) Render(ctx context.Context, w workloads.InstantiatedWorkload) 
 	name := ""
 	port := int32(80)
 	for _, res := range resources {
-		if res.Type != "kubernetes" {
+		if res.Type != workloads.ResourceKindKubernetes {
 			// Not a kubernetes resource
 			continue
 		}
