+++ pkg/workloads/resource_kinds.go
@@ -1,0 +1,14 @@
+// ------------------------------------------------------------
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+// ------------------------------------------------------------
+
+package workloads
+
+const ResourceKindKubernetes = "kubernetes"
+const ResourceKindDaprStateStoreAzureStorage = "dapr.statestore.azurestorage"
+const ResourceKindDaprPubSubTopicAzureServiceBus = "dapr.pubsubtopic.azureservicebus"
+const ResourceKindAzureCosmosDocumentDB = "azure.cosmos.documentdb"
+const ResourceKindAzureServiceBusQueue = "azure.servicebus.queue"
+const ResourceKindAzureKeyVault = "azure.keyvault"
+const ResourceKindAzurePodIdentity = "azure.aadpodidentity"
